Split UploadApkTask into UploadArtifactTask
